pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "txt2stix"
version = "0.0.1b5"
authors = [{ name = "DOGESEC", email = "support@dogesec.com" }]
description = "txt2stix is a Python script that is designed to identify and extract IoCs and TTPs from text files, identify the relationships between them, convert them to STIX 2.1 objects, and output as a STIX 2.1 bundle."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]



dependencies = [
  "pathvalidate>=3.2.0",
  "phonenumbers>=8.13.39",
  "requests==2.32.3",
  "python-dotenv>=1.0.1",
  "schwifty>=2024.6.1",
  "stix2extensions @ https://github.com/muchdogesec/stix2extensions/archive/main.zip",
  "tld>=0.13",
  "tldextract>=5.1.2",
  "validators>=0.28.3",
  "base58>=2.1.1",
  "llama-index-core==0.11.22",
  'llama-index-llms-anthropic==0.3.9',
  'llama-index-llms-gemini==0.3.7',
  'llama-index-llms-openai==0.2.16;',
]

[tool.hatch.metadata]
allow-direct-references = true


[project.urls]
Homepage = "https://github.com/muchdogesec/txt2stix"
Issues = "https://github.com/muchdogesec/txt2stix/issues"


[project.scripts]
stix2arango = "txt2stix.txt2stix:main"

[tool.hatch.build.targets.wheel.force-include]
"includes" = "txt2stix/includes"